Setup seekbar to update its value with song

+++ app/src/main/java/com/hpixel/dreamusicplayer/model/Song.kt
@@ -12,7 +12,7 @@ data class Song(
         var title : String = "no title",
         var filePath : String = "/",
         var displayName : String = "no display",
-        var duration : Long = 0,
+        var duration : Int = 0,
         var albumID : Int = 0
 )
 {
@@ -30,7 +30,7 @@ data class Song(
 
         val DEFAULT_FILE_PATH = "/"
 
-        fun convertToHumanFormat(duration: Long): String {
+        fun convertToHumanFormat(duration: Int): String {
             val min = duration / 60000
             val sec = String.format("%02d", (duration % 60000 / 1000))
             return  "$min:$sec"
@@ -53,7 +53,7 @@ data class Song(
 		title = data[2]
 		filePath = data[3]
 		displayName = data[4]
-		duration = data[5].toLong()
+		duration = data[5].toInt()
         albumName = data[6]
         albumID = data[7].toInt()
 	}
